fixed the double call of the callback function
When reaching the last page the callback function is called twice.
This fixes https://github.com/lapwinglabs/x-ray/pull/53

Added a test but it’s a bit messy because it’s a negative test and I am
not a javascript dev so I hope someone can refactor it in a more
elegant way.