Add halt callback for early exit of pagination
In my use case I needed to stop pagination of a condition was met.
Specifically once an already scraped data point was reached.

This commit adds the option of an object for pagination with
selector (string) and halt (callback function) properties.

The halt function will continue pagination until it returns true.