Emit error on correct stream
This change fixes one of the issues from #98. `state.stream` can be the wrong stream to emit an error to if no path is passed however `ret` will always be the right stream to emit to. 

This makes it possible to listen for 'error' events although an error is still squelched if it's not listened for.