Fix potential issue when selector is a function following comments on commit e7318d591908d04b439c1ff7e93ff9c4b63a89af
When the selector v is a function, the function next() will most likely be called twice: once because there is no return before calling the function v, and once again when the callback for v is called.

As @bdentino notes, this could cause problems down the line because the next() is called too many times.